+++ apis/QryAsset.go
@@ -10,31 +10,33 @@ import (
 
 func ServeAssetWhole(d IDPAsset, w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
-	// process strategy s ["Core", "Value-added", "Opportunistic"]
+	// process parameters
+	// s in ["All", "Core", "Value-added", "Opportunistic"]
 	s := params["strat"]
 	log.Println("Serve asset(checklist) where strategy is ", s)
-
-	if s == "All" {
-		// push all strategy
-		packet, _ := json.Marshal(d.Data)
-		w.WriteHeader(http.StatusOK)
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(packet)
-	} else {
-		// push specific strategy
-		var sendPacket = []assets{}
-		for _, row := range d.Data {
-			switch {
-			case s == row.Strategy:
-				sendPacket = append(sendPacket, row)
-			default:
-			}
-		}
-
-		packet, _ := json.Marshal(sendPacket)
-		w.WriteHeader(http.StatusOK)
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(packet)
+	qry := ReqIDPAsset{
+		Strategy: s,
 	}
 
+	packet, _ := json.Marshal(procAssetQry(qry, d))
+	w.WriteHeader(http.StatusOK)
+	w.Header().Set("Content-Type", "application/json")
+	w.Write(packet)
+}
+
+func procAssetQry(rq ReqIDPAsset, d IDPAsset) []assets {
+	// reads asset query from ReqIDPAsset struct
+	var sendPacket = []assets{}
+	if rq.Strategy == "All" {
+		return d.Data
+	}
+
+	for _, row := range d.Data {
+		switch {
+		case row.Strategy == rq.Strategy:
+			sendPacket = append(sendPacket, row)
+		default:
+		}
+	}
+	return sendPacket
 }