+++ apis/idpV1Strc.go
@@ -1,0 +1,7 @@
+package apis
+
+type IDPDataSet struct {
+	Asset IDPAsset
+	Debt  IDPDebt
+	Macro IDPMacro
+}